Extract condition modules from the Conditional policy
So they can be used from other policies. Actually, we were already using
the Condition and the Operation module from the Rate Limit policy. We
had a dependency between policies, which is something we want to avoid.
